+++ core/source/core/scala/com/rockymadden/stringmetric/similarity/JaccardMetric.scala
@@ -4,7 +4,7 @@ import com.rockymadden.stringmetric.{ ConfigurableStringMetric, MatchTuple, Stri
 import com.rockymadden.stringmetric.tokenization.NGramTokenizer
 
 /* An implementation of the Jaccard metric. */
-class JaccardMetric extends ConfigurableStringMetric[Double, Int] { this: StringFilter =>
+class JaccardMetric extends ConfigurableStringMetric[Int, Double] { this: StringFilter =>
 	final override def compare(charArray1: Array[Char], charArray2: Array[Char])(implicit n: Int): Option[Double] = {
 		if (n <= 0) throw new IllegalArgumentException("Expected valid n.")
