+++ core/source/core/scala/com/rockymadden/stringmetric/similarity/OverlapMetric.scala
@@ -5,7 +5,7 @@ import com.rockymadden.stringmetric.tokenization.NGramTokenizer
 import scala.math
 
 /* An implementation of the overlap metric. */
-class OverlapMetric extends ConfigurableStringMetric[Double, Int] { this: StringFilter =>
+class OverlapMetric extends ConfigurableStringMetric[Int, Double] { this: StringFilter =>
 	final override def compare(charArray1: Array[Char], charArray2: Array[Char])(implicit n: Int): Option[Double] = {
 		if (n <= 0) throw new IllegalArgumentException("Expected valid n.")
