+++ core/source/core/scala/com/rockymadden/stringmetric/similarity/WeightedLevenshteinMetric.scala
@@ -5,7 +5,7 @@ import scala.math.BigDecimal
 
 /** An implementation of a weighted Levenshtein metric. */
 class WeightedLevenshteinMetric
-	extends ConfigurableStringMetric[Double, (BigDecimal, BigDecimal, BigDecimal)] { this: StringFilter =>
+	extends ConfigurableStringMetric[(BigDecimal, BigDecimal, BigDecimal), Double] { this: StringFilter =>
 
 	/** Options order is delete, insert, then substitute weight. */
 	final override def compare(charArray1: Array[Char], charArray2: Array[Char])
